chore: moves all pydantic logic to trestle_rule
