+++ trestlebot/transformers/trestle_rule.py
@@ -5,9 +5,10 @@
 """Trestle Rule class with pydantic."""
 
 
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Tuple, Union
 
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, Field, ValidationError, validator
+from pydantic.error_wrappers import ErrorDict
 
 from trestlebot import const
 
@@ -108,3 +109,32 @@ def get_default_rule() -> TrestleRule:
             include_controls=[Control(id="example")],
         ),
     )
+
+
+# From https://docs.pydantic.dev/latest/errors/errors/
+def loc_to_dot_sep(loc: Tuple[Union[str, int], ...]) -> str:
+    """Convert a tuple of strings and integers to a dot separated string."""
+    path = ""
+    for i, x in enumerate(loc):
+        if isinstance(x, str):
+            if i > 0:
+                path += "."
+            path += x
+        elif isinstance(x, int):
+            path += f"[{x}]"
+        else:
+            raise TypeError("Unexpected type")
+    return path
+
+
+def convert_errors(e: ValidationError) -> List[ErrorDict]:
+    """
+    Convert pydantic validation errors to a list of dictionaries.
+
+    Note: All validations for rules should be done in the pydantic model and
+    formatted through this function.
+    """
+    new_errors: List[ErrorDict] = e.errors()
+    for error in new_errors:
+        error["loc"] = loc_to_dot_sep(error["loc"])
+    return new_errors
